requirements.txt
dj-database-url==2.2.0
dj-static==0.0.6
Django==4.2.7
django-cors-headers==4.2.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.5
msgpack==1.0.7
Pillow==10.0.1
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.9.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.1
